added logic for editing/deleting users
cleaned up controller and added validation for editing/deleting users based on role

+++ UserService/UserController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Model;
 using Model.DataTransfer;
 using Models;
 using Models.DataTransfer;
@@ -10,6 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace UserService
@@ -35,6 +38,7 @@ namespace UserService
             _logger = logger;
         }
 
+
         [HttpPost("create")]
         public async Task<IActionResult> Create([FromBody] CreateUserDto cud)
         {
@@ -50,59 +54,72 @@ namespace UserService
         {
             var user = await _userManager.FindByNameAsync(userForAuthentication.UserName);
             if (user == null || !await _userManager.CheckPasswordAsync(user, userForAuthentication.Password))
-                return Unauthorized(new AuthResponseDto { IsAuthSuccessful = false, ErrorMessage = "Invalid Authentication" }) ;
+                return Unauthorized("Invalid Authentication");
 
             return Ok(await _logic.LoginUser(user));
         }
 
         [HttpGet]
+        [Authorize(Roles="Admin, League Manager, Head Coach")]
         public async Task<IActionResult> GetUsers()
         {
             return Ok(await _logic.GetUsers());
         }
 
-        /// <summary>
-        /// Get a user by their username
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        //[HttpGet("{username}")]
-        //public async Task<IActionResult> GetUserByUsername(string username)
-        //{
-        //    return _mapper.ConvertUserToUserDto(await _logic.GetUserByUsername(username));
-        //}
+        [HttpGet("{username}")]
+        [Authorize]
+        public async Task<ActionResult<UserDto>> GetUserByUsername(string username)
+        {
+            return _mapper.ConvertUserToUserDto(await _logic.GeUserByUsername(username));
+        }
 
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<UserDto>> GetUser(string id)
         {
             return _mapper.ConvertUserToUserDto(await _logic.GetUserById(id));
         }
 
-        /// <summary>
-        /// Get specified User's Role
-        /// </summary>
-        /// <param name="id"></param>
-        /// <param name="editedUser"></param>
-        /// <returns></returns>
-        //[HttpGet("{id}/Role")]
-        //public async Task<IActionResult> GetUserRole(string id)
-        //{
-        //    return Ok(await _logic.GetUserRole(id));
-        //}
+
+        [HttpGet("{id}/Role")]
+        [Authorize]
+        public async Task<IActionResult> GetUserRole(string id)
+        {
+            return Ok(await _logic.GetUserRole(id));
+        }
 
 
         [HttpPut("edit/{id}")]
-        public async Task<ActionResult<ApplicationUser>> EditUser(string id, EditUserDto editedUser)
+        [Authorize]
+        public async Task<ActionResult<UserDto>> EditUser(string id, [FromBody] EditUserDto editedUser)
         {
-            return await _logic.EditUser(id, editedUser);
+            var claimsIdentity = (ClaimsIdentity)User.Identity;
+            string loggedInUserId = await _logic.GetLoggedInUserId(claimsIdentity);
+
+            if (loggedInUserId != id && await _logic.AllowedToAlterUser(loggedInUserId, id) == false)
+                return Forbid("Not authorized to edit this user.");
+
+            return _mapper.ConvertUserToUserDto(await _logic.EditUser(id, editedUser));
         }
 
 
         [HttpDelete("delete/{id}")]
-        public async Task<ApplicationUser> DeleteUser(string id)
+        [Authorize]
+        public async Task<IActionResult> DeleteUser(string id)
         {
-            _logger.LogInformation("User deleted.");
-            return await _logic.DeleteUser(id);
+            var claimsIdentity = (ClaimsIdentity)User.Identity;
+            string loggedInUserId = await _logic.GetLoggedInUserId(claimsIdentity);
+
+            if (loggedInUserId != id && await _logic.AllowedToAlterUser(loggedInUserId, id) == false)
+                return Forbid("Not authorized to delete this user.");
+
+            bool result = await _logic.DeleteUser(id);
+
+            if (result) return Ok("User deleted successfully");
+            return NotFound("User not found");
         }
+
+
+
     }
 }
